move test and adjust
