+++ pipeline/frontend/yaml/compiler/dag_test.go
@@ -1,0 +1,43 @@
+// Copyright 2023 Woodpecker Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package compiler
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	backend_types "go.woodpecker-ci.org/woodpecker/v2/pipeline/backend/types"
+)
+
+func TestConvertDAGToStages(t *testing.T) {
+	steps := map[string]*dagCompilerStep{
+		"step1": {
+			step:      &backend_types.Step{},
+			dependsOn: []string{"step3"},
+		},
+		"step2": {
+			step:      &backend_types.Step{},
+			dependsOn: []string{"step1"},
+		},
+		"step3": {
+			step:      &backend_types.Step{},
+			dependsOn: []string{"step2"},
+		},
+	}
+
+	_, err := convertDAGToStages(steps, "test")
+
+	assert.ErrorContains(t, err, "cycle detected:")
+}
